1. Changed android app package name 2. Changed google-services.json and GoogleService-Info.plist - To be ready for launching our application, I moved our account to 51foundation account. - So, it was necessary to make firebase project again.
