+++ android/app/google-services.json
@@ -1,27 +1,27 @@
 {
   "project_info": {
-    "project_number": "327219256828",
-    "firebase_url": "https://bridgeme-83475.firebaseio.com",
-    "project_id": "bridgeme-83475",
-    "storage_bucket": "bridgeme-83475.appspot.com"
+    "project_number": "899361660530",
+    "firebase_url": "https://bridgeme-5d002.firebaseio.com",
+    "project_id": "bridgeme-5d002",
+    "storage_bucket": "bridgeme-5d002.appspot.com"
   },
   "client": [
     {
       "client_info": {
-        "mobilesdk_app_id": "1:327219256828:android:16c7027ab8aacc2a",
+        "mobilesdk_app_id": "1:899361660530:android:f2aeba822ae3a516",
         "android_client_info": {
-          "package_name": "com.yoda"
+          "package_name": "com.svbootcamp.bridgeme"
         }
       },
       "oauth_client": [
         {
-          "client_id": "327219256828-agn1g6v3q0j3rhetggah086ik40pdtqv.apps.googleusercontent.com",
+          "client_id": "899361660530-devpkc6sgqbt9l295l1jgjs7fq51ibqf.apps.googleusercontent.com",
           "client_type": 3
         }
       ],
       "api_key": [
         {
-          "current_key": "AIzaSyC8is6y9Emkk9AN0Pi8pwXz26qWk7Eoe6s"
+          "current_key": "AIzaSyDFx357Z3Xb26R1UMnnAQ8fg880bAWVhEg"
         }
       ],
       "services": {
