+++ ios/GoogleService-Info.plist
@@ -7,21 +7,21 @@
 	<key>AD_UNIT_ID_FOR_INTERSTITIAL_TEST</key>
 	<string>ca-app-pub-3940256099942544/4411468910</string>
 	<key>CLIENT_ID</key>
-	<string>327219256828-cc6pbt663ltp1ef2g7t5bmret1tae8ko.apps.googleusercontent.com</string>
+	<string>899361660530-6lud34mr54l3qu8u6a3p8jgf52erlm89.apps.googleusercontent.com</string>
 	<key>REVERSED_CLIENT_ID</key>
-	<string>com.googleusercontent.apps.327219256828-cc6pbt663ltp1ef2g7t5bmret1tae8ko</string>
+	<string>com.googleusercontent.apps.899361660530-6lud34mr54l3qu8u6a3p8jgf52erlm89</string>
 	<key>API_KEY</key>
-	<string>AIzaSyDeWY0v5V_IFq6pwSJn5diN-5m8RpqpuHE</string>
+	<string>AIzaSyArLu8Dt4OrgO1D_PVngVhYfhr19EYfkOU</string>
 	<key>GCM_SENDER_ID</key>
-	<string>327219256828</string>
+	<string>899361660530</string>
 	<key>PLIST_VERSION</key>
 	<string>1</string>
 	<key>BUNDLE_ID</key>
 	<string>com.svbootcamp.bridgeme</string>
 	<key>PROJECT_ID</key>
-	<string>bridgeme-83475</string>
+	<string>bridgeme-5d002</string>
 	<key>STORAGE_BUCKET</key>
-	<string>bridgeme-83475.appspot.com</string>
+	<string>bridgeme-5d002.appspot.com</string>
 	<key>IS_ADS_ENABLED</key>
 	<true/>
 	<key>IS_ANALYTICS_ENABLED</key>
@@ -33,8 +33,8 @@
 	<key>IS_SIGNIN_ENABLED</key>
 	<true/>
 	<key>GOOGLE_APP_ID</key>
-	<string>1:327219256828:ios:f2aeba822ae3a516</string>
+	<string>1:899361660530:ios:f2aeba822ae3a516</string>
 	<key>DATABASE_URL</key>
-	<string>https://bridgeme-83475.firebaseio.com</string>
+	<string>https://bridgeme-5d002.firebaseio.com</string>
 </dict>
 </plist>
